refactor(desktop): extract computer-use helpers from main.mjs
The computer-use section (helper app/binary resolution, --check permission
probe, running-app listing for @App mentions, permission-setup launcher)
moves verbatim to electron/computer-use.mjs, following the existing
updater.mjs/migration.mjs sibling-module pattern. Only the IPC registry
consumed these functions; main.mjs imports the four entry points.

Tested:
- node --check both files; typecheck:electron clean; check:electron 50
  methods; desktop node:test 18 pass
- LIVE: booted the Electron dev app — checkComputerUsePermissions (typed
  shape), getComputerUseMcpCommand (array), listRunningApps (2 real apps),
  and workspaceBootstrap (9 workspaces) all roundtrip through the new
  module via the real bridge.
